Move getTextWidth to stepper
