+++ client/landing/stepper/declarative-flow/internals/steps-repository/podcast-title/index.tsx
@@ -9,9 +9,9 @@ import { useState, useRef, useEffect } from 'react';
 import FormLabel from 'calypso/components/forms/form-label';
 import FormSettingExplanation from 'calypso/components/forms/form-setting-explanation';
 import FormInput from 'calypso/components/forms/form-text-input';
-import getTextWidth from 'calypso/landing/gutenboarding/onboarding-block/acquire-intent/get-text-width';
 import usePodcastTitle from 'calypso/landing/stepper/hooks/use-podcast-title';
 import { ONBOARD_STORE, USER_STORE } from 'calypso/landing/stepper/stores';
+import getTextWidth from 'calypso/landing/stepper/utils/get-text-width';
 import { recordTracksEvent } from 'calypso/lib/analytics/tracks';
 import { tip } from 'calypso/signup/icons';
 import type { Step } from '../../types';
